+++ test/observability/ObsHelper.hs
@@ -52,6 +52,7 @@ baseCfg = let secret = encodeUtf8 "reallyreallyreallyreallyverysafe" in
   , configDbSchemas                 = fromList ["test"]
   , configDbConfig                  = False
   , configDbPreConfig               = Nothing
+  , configDbTimezoneEnabled         = False
   , configDbUri                     = "postgresql://"
   , configFilePath                  = Nothing
   , configJWKS                      = rightToMaybe $ parseSecret secret